Fixed the drawing of the Custom HUD buttons
The list of buttons used to draw the menu was CustomHUD.Items, though that includes social_menu that is used for something else. This resulted in an invisible button being made to turn on and off the non-existent social_menu button, while the inputs boxes of the speedometer and the timer were moved in a box below the supposed one.